Adding to Reproduction Logs
System: macOS Ventura Version 13.5 through zsh terminal.

Recommended Fixes for mac Users: Confirming that Home Brew and Python3 is already installed either using the terminal or otherwise. 

General recommendations: The files in the Retrieval Exercise downloaded through the terminal nearly total to 2.46 GB and hence the File Path should be carefully remembered and stored in case storage issues arise in the future.

Status: Everything went smoothly and I was able to reach till the end of the exercise.